Add player type and standardise the game start code
