+++ Engine/src/Civilization.cs
@@ -44,5 +44,15 @@ namespace Civ2engine
         public List<Unit> Units { get; } = new();
 
         public List<City> Cities { get; } = new();
+        
+        public PlayerType PlayerType { get; set; }
+    }
+
+    public enum PlayerType
+    {
+        AI,
+        Local,
+        Remote,
+        Barbarians
     }
 }
