+++ Engine/src/Game.ActionsUnits.cs
@@ -54,7 +54,7 @@ namespace Civ2engine
             // End turn if no units awaiting orders
             if (nextUnit == null && !startOfTurn)
             {
-                if (Options.AlwaysWaitAtEndOfTurn && _activeCiv == _playerCiv)
+                if (Options.AlwaysWaitAtEndOfTurn && _activeCiv.PlayerType != PlayerType.AI)
                 {
                     OnMapEvent?.Invoke(null, new MapEventArgs(MapEventType.WaitAtEndOfTurn));
                 }
